feat(upgrade): 迁移到 Astro V6
