+++ package.json
@@ -16,32 +16,32 @@
     "preinstall": "npx only-allow pnpm"
   },
   "dependencies": {
-    "@astrojs/check": "^0.9.4",
-    "@astrojs/rss": "^4.0.12",
-    "@astrojs/sitemap": "^3.4.2",
-    "@astrojs/svelte": "7.1.0",
+    "@astrojs/check": "^0.9.8",
+    "@astrojs/rss": "^4.0.18",
+    "@astrojs/sitemap": "^3.7.2",
+    "@astrojs/svelte": "8.0.5",
     "@astrojs/tailwind": "^6.0.2",
-    "@expressive-code/core": "^0.41.3",
-    "@expressive-code/plugin-collapsible-sections": "^0.41.3",
-    "@expressive-code/plugin-line-numbers": "^0.41.3",
+    "@expressive-code/core": "^0.41.7",
+    "@expressive-code/plugin-collapsible-sections": "^0.41.7",
+    "@expressive-code/plugin-line-numbers": "^0.41.7",
     "@iconify-json/fa6-brands": "^1.2.6",
     "@iconify-json/fa6-regular": "^1.2.4",
     "@iconify-json/fa6-solid": "^1.2.4",
-    "@iconify-json/ic": "^1.2.2",
-    "@iconify-json/material-symbols": "^1.2.30",
+    "@iconify-json/ic": "^1.2.4",
+    "@iconify-json/material-symbols": "^1.2.67",
     "@iconify/svelte": "^4.2.0",
-    "@swup/astro": "^1.7.0",
-    "@tailwindcss/typography": "^0.5.16",
-    "astro": "5.12.8",
-    "astro-expressive-code": "^0.41.3",
+    "@swup/astro": "^1.8.0",
+    "@tailwindcss/typography": "^0.5.19",
+    "astro": "6.1.8",
+    "astro-expressive-code": "^0.41.7",
     "astro-icon": "^1.1.5",
     "hastscript": "^9.0.1",
-    "katex": "^0.16.22",
-    "markdown-it": "^14.1.0",
+    "katex": "^0.16.45",
+    "markdown-it": "^14.1.1",
     "mdast-util-to-string": "^4.0.0",
-    "node-html-parser": "^7.0.1",
-    "overlayscrollbars": "^2.11.4",
-    "pagefind": "^1.3.0",
+    "node-html-parser": "^7.1.0",
+    "overlayscrollbars": "^2.15.1",
+    "pagefind": "^1.5.2",
     "photoswipe": "^5.4.4",
     "reading-time": "^1.5.0",
     "rehype-autolink-headings": "^7.1.0",
@@ -54,22 +54,22 @@
     "remark-github-admonitions-to-directives": "^1.0.5",
     "remark-math": "^6.0.0",
     "remark-sectionize": "^2.1.0",
-    "sanitize-html": "^2.17.0",
-    "sharp": "^0.34.3",
+    "sanitize-html": "^2.17.3",
+    "sharp": "^0.34.5",
     "stylus": "^0.64.0",
-    "svelte": "^5.37.3",
-    "tailwindcss": "^3.4.17",
-    "typescript": "^5.9.2",
-    "unist-util-visit": "^5.0.0"
+    "svelte": "^5.55.4",
+    "tailwindcss": "^3.4.19",
+    "typescript": "^5.9.3",
+    "unist-util-visit": "^5.1.0"
   },
   "devDependencies": {
-    "@astrojs/ts-plugin": "^1.10.4",
+    "@astrojs/ts-plugin": "^1.10.7",
     "@biomejs/biome": "2.1.3",
     "@rollup/plugin-yaml": "^4.1.2",
     "@types/hast": "^3.0.4",
     "@types/markdown-it": "^14.1.2",
     "@types/mdast": "^4.0.4",
-    "@types/sanitize-html": "^2.16.0",
+    "@types/sanitize-html": "^2.16.1",
     "postcss-import": "^16.1.1",
     "postcss-nesting": "^13.0.2"
   },
