+++ src/components/PostPage.astro
@@ -17,7 +17,7 @@ const interval = 50;
                 category={entry.data.category}
                 published={entry.data.published}
                 updated={entry.data.updated}
-                url={getPostUrlBySlug(entry.slug)}
+                url={getPostUrlBySlug(entry.id)}
                 image={entry.data.image}
                 description={entry.data.description}
                 draft={entry.data.draft}
